Refactor home.vue so the data is extracted by specialized modules, add tests for it
