+++ test/unit/specs/mavlinklogs.spec.js
@@ -1,5 +1,6 @@
 require('mavlink_common_v1.0/mavlink')
 let mavlinkparser = require('../../../src/tools/parsers/mavlinkParser.js')
+import {MavlinkDataExtractor} from '../../../src/tools/mavlinkDataExtractor'
 var glob = require('glob')
 
 // options is optional
@@ -15,7 +16,19 @@ describe('parse tlogs', () => {
         }
         self.postMessage = function (a) {
         }
-        const messageTypes = parser.processData(logfile)
+        const result = parser.processData(logfile)
+        const messageTypes = result.types
+        const messages = result.messages
         expect(Object.keys(messageTypes)).toContain('HEARTBEAT')
+        expect(MavlinkDataExtractor.extractArmedEvents(messages).length).toBeGreaterThan(0)
+        expect(Object.keys(MavlinkDataExtractor.extractAttitudes(messages)).length).toBeGreaterThan(100)
+        expect(MavlinkDataExtractor.extractFlightModes(messages).length).toBeGreaterThan(0)
+
+        expect(MavlinkDataExtractor.extractParams(messages)).toBeDefined()
+
+        expect(MavlinkDataExtractor.extractTextMessages(messages).length).toBeGreaterThan(0)
+        let trajectory = MavlinkDataExtractor.extractTrajectory(messages)
+        expect(Object.keys(trajectory).length).toBeGreaterThan(1)
+        expect(trajectory.trajectory.length).toBeGreaterThan(100)
     })
 })
